+++ MLX90393Hal.h
@@ -19,8 +19,8 @@ class MLX90393Hal {
         virtual bool has_drdy_pin() {return this->drdy_pin_ != -1;}
         virtual bool transceive(const uint8_t *request, size_t request_size, uint8_t *response, size_t response_size);
         virtual bool read_drdy_pin();
-        virtual void sleep_millis(long millis);
-        virtual void sleep_micros(long micros);
+        virtual void sleep_millis(uint32_t millis);
+        virtual void sleep_micros(uint32_t micros);
     protected:
         uint8_t I2C_address_ = 0;
         int drdy_pin_ = -1;
